mount /mnt

+++ executor/invoker.py
@@ -802,14 +802,14 @@ def vMDiskSnapshotWatcher(group=GROUP_VM_DISK_SNAPSHOT, version=VERSION_VM_DISK_
                     status = 'Done(Success)'
                     if _isDeleteDiskExternalSnapshot(the_cmd_key):
                         if data:
-                            if data['delete_ss']:
+                            if 'delete_ss' in data.keys() and data['delete_ss']:
                                 for delete_ss in data['delete_ss']:
                                     try:
                                         deleteStructure(delete_ss, V1DeleteOptions(), group, version, plural)
                                     except ApiException, e:
                                         if e.reason == 'Not Found':
                                             logger.debug('**Object %s already deleted.' % delete_ss)
-                            if data['need_to_modify']:
+                            if 'need_to_modify' in data.keys() and data['need_to_modify']:
                                 try:
                                     modify_snapshot(data['pool'], data['disk'], data['need_to_modify'], group, version, plural)
                                 except Exception:
